cleanup/code review
trying to make searching cleaner. currently, pressing button doesn't work. queries an object instead of the search term. one option is to create a state to track the current search term (on change, and set inputvalue). But this seems inefficient givent that there is already a searchTerm variable. However, search term is currently responsible for fetching data when changed (useEffect in Serach context), so it may be ok to wait for the search term with the new variable, and onclick/onsearch, set it to the searchTerm