Accidentally altered all vet data, made all firstnames the same
Forgot to add where clause in the sql command, caused all doctor's to be renamed to abc. oops
Noticed a small bug. After saving data in personal data, and navigating to dashboard, the name/lastname doesn't change in the dashboard. This is because dashboard data is being pulled seperately from personalData. Might not be relevenat in the future since the dashboard will not have user info, dashboard data will contain compeltely different things
-Started adding new user only permissions in register function. will need to check for this cookie in newDoctor/Patient functions. If exists & doctoruuid/verifyied , allow to fill out. if verified, but newDoctor doesn't exist, navigate to personal data. If neither exists, or if newDoctor exists but uuid/verify doesn't, navigate to register.
on navigating to the dashboard after newDoctor, will need to clear the cookie. Do on the back-end. to prevent people from re-entering