Description not properly saving/encrypting/decrypting
Not sure what the problem is. Whenever description is saved, it is either saving as a buffer, or not being read properly. Could be because of the Blob datatype. Consider changing blob to VARCHAR, and then limiting the number of characters the user can input on server-side. Made description into a useState empty object, will need to figure out how to not make [Object Object] show when description is queried.