Documentation leading to code changes
Beginning to re-read/document back-end code, and found an inefficiency. No point in returning the tokenvalue, can just return the type. On server-side, had to change any path that used user_verification.
There is a small bug. When a user is logged in, and they navigate to new-patient/doctor, their name disappears from the header. When re-loading, re-appears. Very strange.