beginning education integration
Education is a prickly pear. Education is not just 1 entry like in the insurance case, it is an object of IDs, each of which must be compared against the correct entry in the table. Feeding preVet and vet education into the same function, since the logic is almost identical. pre-vet just has a major that vet doesn't.