Major changes to locations
1. Updated the phone_numbers sql table such that when an address_ID from the addresses table that is linked to an address in phone_numbers is deleted, the phone_numbers record is also delted. Used an on delete cascasde.
2. Adding compeltely new location data (when there has never been previously added Data) to to the database works.
To-do: make adding new data once data already exists work, same for updating and deleting data (and if there are mulitple records that need to be updated/deleted (for loop)).
Will need to send back the correct Address/phone data to the fornt-end with the corect replaced address_ID when saving. or figure out another way to selectively update the address_IDs of the changed data. For completely new data, can send back the data just entered into the DB. for existing data, can send back a contcatenation of updated and addedData.